+++ src/main/java/insane96mcp/iguanatweaksreborn/module/experience/enchantments/enchantment/protection/OverallProtection.java
@@ -22,6 +22,11 @@ public class OverallProtection extends ITRProtectionEnchantment {
     }
 
     @Override
+    public boolean isTreasureOnly() {
+        return true;
+    }
+
+    @Override
     public float getDamageReductionPerLevel() {
         return 0.06f;
     }